Fixed a few tests.| #2385
