+++ FlowCrypt/src/test/java/com/flowcrypt/email/security/pgp/PgpMsgTest.kt
@@ -316,6 +316,7 @@ class PgpMsgTest {
   // -------------------------------------------------------------------------------------------
 
   @Test
+
   @Ignore("Should be reworked after switching to a new logic that uses RawBlockParser")
   fun multipleComplexMessagesTest() {
     val testFiles = listOf(
@@ -341,6 +342,7 @@ class PgpMsgTest {
 
   // Use this one for debugging
   @Test
+
   @Ignore("Should be reworked after switching to a new logic that uses RawBlockParser")
   fun singleComplexMessageTest() {
     val testFile = "verify - Kraken - urldecode signature-0.json"
